src/requirements.txt
aiomysql==0.1.1
aioredis==2.0.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
bcrypt==3.2.2
cffi==1.15.0
click==8.1.3
cryptography==37.0.2
databases==0.6.0
ecdsa==0.17.0
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
orjson==3.7.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
PyMySQL==1.0.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.37
starlette==0.19.1
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
